pyproject.toml
[project]
name = "rag-retrieval-benchmark"
version = "0.1.0"
description = "Comprehensive RAG retrieval methods benchmark on T²-RAGBench"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Süleyman Akarsu"}]

dependencies = [
    # Data
    "datasets>=3.0.0",
    "pandas>=2.2.0",
    "pyarrow>=17.0.0",

    # Sparse Retrieval
    "rank-bm25>=0.2.2",

    # Dense Retrieval
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0,<2.0.0",

    # ColBERT
    "ragatouille>=0.0.9",

    # API Clients
    "openai>=1.40.0",
    "cohere>=5.0.0",
    "voyageai>=0.3.0",

    # Reranking
    "flashrank>=0.2.0",

    # Generation & Evaluation
    "rouge-score>=0.1.2",
    "bert-score>=0.3.13",
    "ragas>=0.2.0",
    "nltk>=3.9.0",
    "evaluate>=0.4.0",
    "scipy>=1.13.0",

    # Config & Utils
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "click>=8.1.0",

    # Tokenization
    "tiktoken>=0.7.0",
    "transformers>=4.44.0",
    "torch>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
]
graph = [
    "lightrag-hku>=0.1.0",
    "networkx>=3.3",
    "graspologic>=3.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
